refactor(delivery): split durable queue owners and remove cycle
