+++ src/cron/isolated-agent/delivery-dispatch-types.ts
@@ -1,0 +1,58 @@
+import type { ReplyPayload } from "../../auto-reply/reply-payload.js";
+import type { CliDeps } from "../../cli/outbound-send-deps.js";
+import type { OpenClawConfig } from "../../config/types.openclaw.js";
+import type { TtsAutoMode } from "../../config/types.tts.js";
+import type { SourceDeliveryOutcome } from "../../infra/outbound/source-delivery-plan.js";
+import type { CronJob, CronRunTelemetry } from "../types.js";
+import type { DeliveryTargetResolution } from "./delivery-target.js";
+import type { RunCronAgentTurnResult } from "./run.types.js";
+
+export type SuccessfulCronDeliveryTarget = Extract<DeliveryTargetResolution, { ok: true }>;
+
+export type DispatchCronDeliveryParams = {
+  cfg: OpenClawConfig;
+  cfgWithAgentDefaults: OpenClawConfig;
+  deps: CliDeps;
+  job: CronJob;
+  agentId: string;
+  agentSessionKey: string;
+  runSessionKey: string;
+  sessionId: string;
+  lifecycleRevision: string;
+  sessionUpdatedAt: number;
+  beforeSessionDelete?: () => void;
+  runStartedAt: number;
+  runEndedAt: number;
+  timeoutMs: number;
+  resolvedDelivery: DeliveryTargetResolution;
+  deliveryRequested: boolean;
+  skipHeartbeatDelivery: boolean;
+  sourceDeliveryOutcome: SourceDeliveryOutcome;
+  deliveryBestEffort: boolean;
+  deliveryPayloadHasStructuredContent: boolean;
+  deliveryPayloads: ReplyPayload[];
+  synthesizedText?: string;
+  ttsAuto?: TtsAutoMode;
+  summary?: string;
+  outputText?: string;
+  telemetry?: CronRunTelemetry;
+  abortSignal?: AbortSignal;
+  isAborted: () => boolean;
+  abortReason: () => string;
+  withRunSession: (
+    result: Omit<RunCronAgentTurnResult, "sessionId" | "sessionKey">,
+  ) => RunCronAgentTurnResult;
+};
+
+/** Mutable delivery-dispatch accumulator returned to the isolated cron runner. */
+export type DispatchCronDeliveryState = {
+  result?: RunCronAgentTurnResult;
+  delivered: boolean;
+  deliveryAttempted: boolean;
+  deliveryError?: string;
+  cronRunSessionCleanupAttempted: boolean;
+  summary?: string;
+  outputText?: string;
+  synthesizedText?: string;
+  deliveryPayloads: ReplyPayload[];
+};
